6610 move additional group keys out of filesearch class
