+++ Core/src/org/sleuthkit/autopsy/discovery/search/Bundle.properties-MERGED
@@ -1,0 +1,115 @@
+AttributeSearchData.AttributeType.Domain.displayName=Domain
+AttributeSearchData.AttributeType.Other.displayName=Other
+DiscoveryKeyUtils.FileTagGroupKey.noSets=None
+DiscoveryKeyUtils.NoGroupingGroupKey.allFiles=All Files
+FileGroup.groupSortingAlgorithm.groupName.text=Group Name
+FileGroup.groupSortingAlgorithm.groupSize.text=Group Size
+# {0} - Data source name
+# {1} - Data source ID
+FileSearch.DataSourceGroupKey.datasourceAndID={0}(ID: {1})
+# {0} - Data source ID
+FileSearch.DataSourceGroupKey.idOnly=Data source (ID: {0})
+FileSearch.documentSummary.noBytes=No bytes read for document, unable to display preview.
+FileSearch.documentSummary.noPreview=No preview available.
+FileSearch.GroupingAttributeType.datasource.displayName=Data Source
+FileSearch.GroupingAttributeType.fileType.displayName=File Type
+FileSearch.GroupingAttributeType.frequency.displayName=Past Occurrences
+FileSearch.GroupingAttributeType.hash.displayName=Hash Set
+FileSearch.GroupingAttributeType.interestingItem.displayName=Interesting Item
+FileSearch.GroupingAttributeType.keywordList.displayName=Keyword
+FileSearch.GroupingAttributeType.none.displayName=None
+FileSearch.GroupingAttributeType.object.displayName=Object Detected
+FileSearch.GroupingAttributeType.parent.displayName=Parent Folder
+FileSearch.GroupingAttributeType.size.displayName=File Size
+FileSearch.GroupingAttributeType.tag.displayName=Tag
+FileSearch.HashHitsGroupKey.noHashHits=None
+FileSearch.InterestingItemGroupKey.noSets=None
+FileSearch.KeywordListGroupKey.noKeywords=None
+FileSearch.ObjectDetectedGroupKey.noSets=None
+FileSearchData.FileSize.100kbto1mb=: 100KB-1MB
+FileSearchData.FileSize.100mbto1gb=: 100MB-1GB
+FileSearchData.FileSize.10PlusGb=: 10GB+
+FileSearchData.FileSize.16kbto100kb=: 16-100KB
+FileSearchData.FileSize.1gbto5gb=: 1-5GB
+FileSearchData.FileSize.1mbto50mb=: 1-50MB
+FileSearchData.FileSize.200PlusMb=: 200MB+
+FileSearchData.FileSize.500kbto100mb=: 500KB-100MB
+FileSearchData.FileSize.50mbto200mb=: 50-200MB
+FileSearchData.FileSize.5gbto10gb=: 5-10GB
+FileSearchData.FileSize.LARGE.displayName=Large
+FileSearchData.FileSize.MEDIUM.displayName=Medium
+FileSearchData.FileSize.SMALL.displayName=Small
+FileSearchData.FileSize.upTo16kb=: 0-16KB
+FileSearchData.FileSize.upTo500kb=: 0-500KB
+FileSearchData.FileSize.XLARGE.displayName=XLarge
+FileSearchData.FileSize.XSMALL.displayName=XSmall
+FileSearchData.FileSize.XXLARGE.displayName=XXLarge
+FileSearchData.FileType.Audio.displayName=Audio
+FileSearchData.FileType.Documents.displayName=Documents
+FileSearchData.FileType.Executables.displayName=Executables
+FileSearchData.FileType.Image.displayName=Image
+FileSearchData.FileType.Other.displayName=Other/Unknown
+FileSearchData.FileType.Video.displayName=Video
+FileSearchData.Frequency.common.displayName=Common (11 - 100)
+FileSearchData.Frequency.known.displayName=Known (NSRL)
+FileSearchData.Frequency.rare.displayName=Rare (2-10)
+FileSearchData.Frequency.unique.displayName=Unique (1)
+FileSearchData.Frequency.unknown.displayName=Unknown
+FileSearchData.Frequency.verycommon.displayName=Very Common (100+)
+FileSearchData.Score.interesting.displayName=Interesting
+FileSearchData.Score.notable.displayName=Notable
+FileSearchData.Score.unknown.displayName=Unknown
+FileSearchFiltering.concatenateSetNamesForDisplay.comma=, 
+# {0} - filters
+FileSearchFiltering.HashSetFilter.desc=Hash set hits in set(s): {0}
+FileSearchFiltering.KnownFilter.desc=which are not known
+FileSearchFiltering.PreviouslyNotableFilter.desc=that were previously marked as notable
+# {0} - tag names
+FileSearchFiltering.TagsFilter.desc=Tagged {0}
+FileSearchFiltering.TagsFilter.or=, 
+FileSearchFiltering.UserCreatedFilter.desc=that contain EXIF data
+FileSorter.SortingMethod.datasource.displayName=Data Source
+FileSorter.SortingMethod.filename.displayName=File Name
+FileSorter.SortingMethod.filesize.displayName=File Size
+FileSorter.SortingMethod.filetype.displayName=File Type
+FileSorter.SortingMethod.frequency.displayName=Central Repo Frequency
+FileSorter.SortingMethod.fullPath.displayName=Full Path
+FileSorter.SortingMethod.keywordlist.displayName=Keyword List Names
+ResultFile.score.interestingResult.description=At least one instance of the file has an interesting result associated with it.
+ResultFile.score.notableFile.description=At least one instance of the file was recognized as notable.
+ResultFile.score.notableTaggedFile.description=At least one instance of the file is tagged with a notable tag.
+ResultFile.score.taggedFile.description=At least one instance of the file has been tagged.
+# {0} - Data source name
+# {1} - Data source ID
+SearchFiltering.DataSourceFilter.datasource={0}({1})
+# {0} - filters
+SearchFiltering.DataSourceFilter.desc=Data source(s): {0}
+SearchFiltering.DataSourceFilter.or=, 
+# {0} - filters
+SearchFiltering.FileTypeFilter.desc=Type: {0}
+SearchFiltering.FileTypeFilter.or=, 
+# {0} - filters
+SearchFiltering.FrequencyFilter.desc=Past occurrences: {0}
+SearchFiltering.FrequencyFilter.or=, 
+# {0} - filters
+SearchFiltering.InterestingItemSetFilter.desc=Interesting item hits in set(s): {0}
+# {0} - filters
+SearchFiltering.KeywordListFilter.desc=Keywords in list(s): {0}
+# {0} - filters
+SearchFiltering.ObjectDetectionFilter.desc=Objects detected in set(s): {0}
+# {0} - filters
+SearchFiltering.ParentFilter.desc=Paths matching: {0}
+SearchFiltering.ParentFilter.exact=(exact match)
+SearchFiltering.ParentFilter.excluded=(excluded)
+SearchFiltering.ParentFilter.included=(included)
+SearchFiltering.ParentFilter.or=, 
+SearchFiltering.ParentFilter.substring=(substring)
+SearchFiltering.ParentSearchTerm.excludeString=\ (exclude)
+SearchFiltering.ParentSearchTerm.fullString=\ (exact)
+SearchFiltering.ParentSearchTerm.includeString=\ (include)
+SearchFiltering.ParentSearchTerm.subString=\ (substring)
+# {0} - filters
+SearchFiltering.ScoreFilter.desc=Score(s) of : {0}
+# {0} - filters
+SearchFiltering.SizeFilter.desc=Size(s): {0}
+SearchFiltering.SizeFilter.or=,
