+++ Core/src/org/sleuthkit/autopsy/discovery/search/FileSearch.java
@@ -25,14 +25,12 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.logging.Level;
 import org.apache.commons.lang.StringUtils;
@@ -42,13 +40,11 @@ import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepoException;
 import org.sleuthkit.autopsy.centralrepository.datamodel.CentralRepoDbUtil;
 import org.sleuthkit.autopsy.centralrepository.datamodel.InstanceTableCallback;
 import org.sleuthkit.autopsy.coreutils.Logger;
-import org.sleuthkit.autopsy.discovery.search.FileSearchData.FileType;
 import org.sleuthkit.autopsy.discovery.search.FileSearchData.Frequency;
 import org.sleuthkit.datamodel.AbstractFile;
 import org.sleuthkit.datamodel.BlackboardArtifact;
 import org.sleuthkit.datamodel.BlackboardAttribute;
 import org.sleuthkit.datamodel.CaseDbAccessManager;
-import org.sleuthkit.datamodel.Content;
 import org.sleuthkit.datamodel.ContentTag;
 import org.sleuthkit.datamodel.SleuthkitCase;
 import org.sleuthkit.datamodel.TskCoreException;
@@ -69,7 +65,6 @@ public class FileSearch {
     private static final Cache<SearchKey, Map<GroupKey, List<ResultFile>>> searchCache = CacheBuilder.newBuilder()
             .maximumSize(MAXIMUM_CACHE_SIZE)
             .build();
-
 
     /**
      * Run the file search and returns the SearchResults object for debugging.
@@ -237,7 +232,7 @@ public class FileSearch {
         "FileSearch.documentSummary.noBytes=No bytes read for document, unable to display preview."})
     public static TextSummary summarize(AbstractFile file) {
         TextSummary summary = null;
-        TextSummarizer localSummarizer; 
+        TextSummarizer localSummarizer;
         synchronized (searchCache) {
             localSummarizer = SummaryHelpers.getLocalSummarizer();
 
@@ -417,16 +412,14 @@ public class FileSearch {
         }
     }
 
-   
-
     /**
      * Attribute for grouping/sorting by file size
      */
-    static class FileSizeAttribute extends AttributeType {
+    public static class FileSizeAttribute extends AttributeType {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new FileSizeGroupKey(file);
+            return new DiscoveryKeyUtils.FileSizeGroupKey(file);
         }
     }
 
@@ -437,115 +430,18 @@ public class FileSearch {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new ParentPathGroupKey(file);
+            return new DiscoveryKeyUtils.ParentPathGroupKey(file);
         }
     }
 
     /**
-     * Key representing a parent path group
+     * Default attribute used to make one group
      */
-    private static class ParentPathGroupKey extends GroupKey {
-
-        private String parentPath;
-        private Long parentID;
-
-        ParentPathGroupKey(ResultFile file) {
-            Content parent;
-            try {
-                parent = file.getFirstInstance().getParent();
-            } catch (TskCoreException ignored) {
-                parent = null;
-            }
-            //Find the directory this file is in if it is an embedded file
-            while (parent != null && parent instanceof AbstractFile && ((AbstractFile) parent).isFile()) {
-                try {
-                    parent = parent.getParent();
-                } catch (TskCoreException ignored) {
-                    parent = null;
-                }
-            }
-            setParentPathAndID(parent, file);
-        }
-
-        /**
-         * Helper method to set the parent path and parent ID.
-         *
-         * @param parent The parent content object.
-         * @param file   The ResultFile object.
-         */
-        private void setParentPathAndID(Content parent, ResultFile file) {
-            if (parent != null) {
-                try {
-                    parentPath = parent.getUniquePath();
-                    parentID = parent.getId();
-                } catch (TskCoreException ignored) {
-                    //catch block left blank purposefully next if statement will handle case when exception takes place as well as when parent is null
-                }
-
-            }
-            if (parentPath == null) {
-                if (file.getFirstInstance().getParentPath() != null) {
-                    parentPath = file.getFirstInstance().getParentPath();
-                } else {
-                    parentPath = ""; // NON-NLS
-                }
-                parentID = -1L;
-            }
-        }
+    static class NoGroupingAttribute extends FileSearch.AttributeType {
 
         @Override
-        String getDisplayName() {
-            return getParentPath();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof ParentPathGroupKey) {
-                ParentPathGroupKey otherParentPathGroupKey = (ParentPathGroupKey) otherGroupKey;
-                int comparisonResult = getParentPath().compareTo(otherParentPathGroupKey.getParentPath());
-                if (comparisonResult == 0) {
-                    comparisonResult = getParentID().compareTo(otherParentPathGroupKey.getParentID());
-                }
-                return comparisonResult;
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof ParentPathGroupKey)) {
-                return false;
-            }
-
-            ParentPathGroupKey otherParentPathGroupKey = (ParentPathGroupKey) otherKey;
-            return getParentPath().equals(otherParentPathGroupKey.getParentPath()) && getParentID().equals(otherParentPathGroupKey.getParentID());
-        }
-
-        @Override
-        public int hashCode() {
-            int hashCode = 11;
-            hashCode = 61 * hashCode + Objects.hash(getParentPath());
-            hashCode = 61 * hashCode + Objects.hash(getParentID());
-            return hashCode;
-        }
-
-        /**
-         * @return the parentPath
-         */
-        String getParentPath() {
-            return parentPath;
-        }
-
-        /**
-         * @return the parentID
-         */
-        Long getParentID() {
-            return parentID;
+        public GroupKey getGroupKey(ResultFile file) {
+            return new DiscoveryKeyUtils.NoGroupingGroupKey();
         }
     }
 
@@ -556,76 +452,7 @@ public class FileSearch {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new DataSourceGroupKey(file);
-        }
-    }
-
-    /**
-     * Key representing a data source group
-     */
-    private static class DataSourceGroupKey extends GroupKey {
-
-        private final long dataSourceID;
-        private String displayName;
-
-        @NbBundle.Messages({
-            "# {0} - Data source name",
-            "# {1} - Data source ID",
-            "FileSearch.DataSourceGroupKey.datasourceAndID={0}(ID: {1})",
-            "# {0} - Data source ID",
-            "FileSearch.DataSourceGroupKey.idOnly=Data source (ID: {0})"})
-        DataSourceGroupKey(ResultFile file) {
-            dataSourceID = file.getFirstInstance().getDataSourceObjectId();
-
-            try {
-                // The data source should be cached so this won't actually be a database query.
-                Content ds = file.getFirstInstance().getDataSource();
-                displayName = Bundle.FileSearch_DataSourceGroupKey_datasourceAndID(ds.getName(), ds.getId());
-            } catch (TskCoreException ex) {
-                logger.log(Level.WARNING, "Error looking up data source with ID " + dataSourceID, ex); // NON-NLS
-                displayName = Bundle.FileSearch_DataSourceGroupKey_idOnly(dataSourceID);
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return displayName;
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof DataSourceGroupKey) {
-                DataSourceGroupKey otherDataSourceGroupKey = (DataSourceGroupKey) otherGroupKey;
-                return Long.compare(getDataSourceID(), otherDataSourceGroupKey.getDataSourceID());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof DataSourceGroupKey)) {
-                return false;
-            }
-
-            DataSourceGroupKey otherDataSourceGroupKey = (DataSourceGroupKey) otherKey;
-            return getDataSourceID() == otherDataSourceGroupKey.getDataSourceID();
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getDataSourceID());
-        }
-
-        /**
-         * @return the dataSourceID
-         */
-        long getDataSourceID() {
-            return dataSourceID;
+            return new DiscoveryKeyUtils.DataSourceGroupKey(file);
         }
     }
 
@@ -636,60 +463,7 @@ public class FileSearch {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new FileTypeGroupKey(file);
-        }
-    }
-
-    /**
-     * Key representing a file type group
-     */
-    private static class FileTypeGroupKey extends GroupKey {
-
-        private final FileType fileType;
-
-        FileTypeGroupKey(ResultFile file) {
-            fileType = file.getFileType();
-        }
-
-        @Override
-        String getDisplayName() {
-            return getFileType().toString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof FileTypeGroupKey) {
-                FileTypeGroupKey otherFileTypeGroupKey = (FileTypeGroupKey) otherGroupKey;
-                return Integer.compare(getFileType().getRanking(), otherFileTypeGroupKey.getFileType().getRanking());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof FileTypeGroupKey)) {
-                return false;
-            }
-
-            FileTypeGroupKey otherFileTypeGroupKey = (FileTypeGroupKey) otherKey;
-            return getFileType().equals(otherFileTypeGroupKey.getFileType());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getFileType().getRanking());
-        }
-
-        /**
-         * @return the fileType
-         */
-        FileType getFileType() {
-            return fileType;
+            return new DiscoveryKeyUtils.FileTypeGroupKey(file);
         }
     }
 
@@ -700,7 +474,7 @@ public class FileSearch {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new KeywordListGroupKey(file);
+            return new DiscoveryKeyUtils.KeywordListGroupKey(file);
         }
 
         @Override
@@ -766,87 +540,6 @@ public class FileSearch {
     }
 
     /**
-     * Key representing a keyword list group
-     */
-    private static class KeywordListGroupKey extends GroupKey {
-
-        private final List<String> keywordListNames;
-        private final String keywordListNamesString;
-
-        @NbBundle.Messages({
-            "FileSearch.KeywordListGroupKey.noKeywords=None"})
-        KeywordListGroupKey(ResultFile file) {
-            keywordListNames = file.getKeywordListNames();
-
-            if (keywordListNames.isEmpty()) {
-                keywordListNamesString = Bundle.FileSearch_KeywordListGroupKey_noKeywords();
-            } else {
-                keywordListNamesString = String.join(",", keywordListNames); // NON-NLS
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getKeywordListNamesString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof KeywordListGroupKey) {
-                KeywordListGroupKey otherKeywordListNamesGroupKey = (KeywordListGroupKey) otherGroupKey;
-
-                // Put the empty list at the end
-                if (getKeywordListNames().isEmpty()) {
-                    if (otherKeywordListNamesGroupKey.getKeywordListNames().isEmpty()) {
-                        return 0;
-                    } else {
-                        return 1;
-                    }
-                } else if (otherKeywordListNamesGroupKey.getKeywordListNames().isEmpty()) {
-                    return -1;
-                }
-
-                return getKeywordListNamesString().compareTo(otherKeywordListNamesGroupKey.getKeywordListNamesString());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof KeywordListGroupKey)) {
-                return false;
-            }
-
-            KeywordListGroupKey otherKeywordListGroupKey = (KeywordListGroupKey) otherKey;
-            return getKeywordListNamesString().equals(otherKeywordListGroupKey.getKeywordListNamesString());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getKeywordListNamesString());
-        }
-
-        /**
-         * @return the keywordListNames
-         */
-        List<String> getKeywordListNames() {
-            return Collections.unmodifiableList(keywordListNames);
-        }
-
-        /**
-         * @return the keywordListNamesString
-         */
-        String getKeywordListNamesString() {
-            return keywordListNamesString;
-        }
-    }
-
-    /**
      * Attribute for grouping/sorting by frequency in the central repository
      */
     static class FrequencyAttribute extends AttributeType {
@@ -855,7 +548,7 @@ public class FileSearch {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new FrequencyGroupKey(file);
+            return new DiscoveryKeyUtils.FrequencyGroupKey(file);
         }
 
         @Override
@@ -944,66 +637,13 @@ public class FileSearch {
     }
 
     /**
-     * Key representing a central repository frequency group
-     */
-    private static class FrequencyGroupKey extends GroupKey {
-
-        private final Frequency frequency;
-
-        FrequencyGroupKey(ResultFile file) {
-            frequency = file.getFrequency();
-        }
-
-        @Override
-        String getDisplayName() {
-            return getFrequency().toString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof FrequencyGroupKey) {
-                FrequencyGroupKey otherFrequencyGroupKey = (FrequencyGroupKey) otherGroupKey;
-                return Integer.compare(getFrequency().getRanking(), otherFrequencyGroupKey.getFrequency().getRanking());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof FrequencyGroupKey)) {
-                return false;
-            }
-
-            FrequencyGroupKey otherFrequencyGroupKey = (FrequencyGroupKey) otherKey;
-            return getFrequency().equals(otherFrequencyGroupKey.getFrequency());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getFrequency().getRanking());
-        }
-
-        /**
-         * @return the frequency
-         */
-        Frequency getFrequency() {
-            return frequency;
-        }
-    }
-
-    /**
      * Attribute for grouping/sorting by hash set lists
      */
     static class HashHitsAttribute extends AttributeType {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new HashHitsGroupKey(file);
+            return new DiscoveryKeyUtils.HashHitsGroupKey(file);
         }
 
         @Override
@@ -1068,94 +708,13 @@ public class FileSearch {
     }
 
     /**
-     * Key representing a hash hits group
-     */
-    private static class HashHitsGroupKey extends GroupKey {
-
-        private final List<String> hashSetNames;
-        private final String hashSetNamesString;
-
-        @NbBundle.Messages({
-            "FileSearch.HashHitsGroupKey.noHashHits=None"})
-        HashHitsGroupKey(ResultFile file) {
-            hashSetNames = file.getHashSetNames();
-
-            if (hashSetNames.isEmpty()) {
-                hashSetNamesString = Bundle.FileSearch_HashHitsGroupKey_noHashHits();
-            } else {
-                hashSetNamesString = String.join(",", hashSetNames); // NON-NLS
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getHashSetNamesString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof HashHitsGroupKey) {
-                HashHitsGroupKey otherHashHitsGroupKey = (HashHitsGroupKey) otherGroupKey;
-
-                // Put the empty list at the end
-                if (getHashSetNames().isEmpty()) {
-                    if (otherHashHitsGroupKey.getHashSetNames().isEmpty()) {
-                        return 0;
-                    } else {
-                        return 1;
-                    }
-                } else if (otherHashHitsGroupKey.getHashSetNames().isEmpty()) {
-                    return -1;
-                }
-
-                return getHashSetNamesString().compareTo(otherHashHitsGroupKey.getHashSetNamesString());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof HashHitsGroupKey)) {
-                return false;
-            }
-
-            HashHitsGroupKey otherHashHitsGroupKey = (HashHitsGroupKey) otherKey;
-            return getHashSetNamesString().equals(otherHashHitsGroupKey.getHashSetNamesString());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getHashSetNamesString());
-        }
-
-        /**
-         * @return the hashSetNames
-         */
-        List<String> getHashSetNames() {
-            return Collections.unmodifiableList(hashSetNames);
-        }
-
-        /**
-         * @return the hashSetNamesString
-         */
-        String getHashSetNamesString() {
-            return hashSetNamesString;
-        }
-    }
-
-    /**
      * Attribute for grouping/sorting by interesting item set lists
      */
     static class InterestingItemAttribute extends AttributeType {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new InterestingItemGroupKey(file);
+            return new DiscoveryKeyUtils.InterestingItemGroupKey(file);
         }
 
         @Override
@@ -1222,94 +781,13 @@ public class FileSearch {
     }
 
     /**
-     * Key representing a interesting item set group
-     */
-    private static class InterestingItemGroupKey extends GroupKey {
-
-        private final List<String> interestingItemSetNames;
-        private final String interestingItemSetNamesString;
-
-        @NbBundle.Messages({
-            "FileSearch.InterestingItemGroupKey.noSets=None"})
-        InterestingItemGroupKey(ResultFile file) {
-            interestingItemSetNames = file.getInterestingSetNames();
-
-            if (interestingItemSetNames.isEmpty()) {
-                interestingItemSetNamesString = Bundle.FileSearch_InterestingItemGroupKey_noSets();
-            } else {
-                interestingItemSetNamesString = String.join(",", interestingItemSetNames); // NON-NLS
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getInterestingItemSetNamesString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof InterestingItemGroupKey) {
-                InterestingItemGroupKey otherInterestingItemGroupKey = (InterestingItemGroupKey) otherGroupKey;
-
-                // Put the empty list at the end
-                if (this.getInterestingItemSetNames().isEmpty()) {
-                    if (otherInterestingItemGroupKey.getInterestingItemSetNames().isEmpty()) {
-                        return 0;
-                    } else {
-                        return 1;
-                    }
-                } else if (otherInterestingItemGroupKey.getInterestingItemSetNames().isEmpty()) {
-                    return -1;
-                }
-
-                return getInterestingItemSetNamesString().compareTo(otherInterestingItemGroupKey.getInterestingItemSetNamesString());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof InterestingItemGroupKey)) {
-                return false;
-            }
-
-            InterestingItemGroupKey otherInterestingItemGroupKey = (InterestingItemGroupKey) otherKey;
-            return getInterestingItemSetNamesString().equals(otherInterestingItemGroupKey.getInterestingItemSetNamesString());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getInterestingItemSetNamesString());
-        }
-
-        /**
-         * @return the interestingItemSetNames
-         */
-        List<String> getInterestingItemSetNames() {
-            return Collections.unmodifiableList(interestingItemSetNames);
-        }
-
-        /**
-         * @return the interestingItemSetNamesString
-         */
-        String getInterestingItemSetNamesString() {
-            return interestingItemSetNamesString;
-        }
-    }
-
-    /**
      * Attribute for grouping/sorting by objects detected
      */
     static class ObjectDetectedAttribute extends AttributeType {
 
         @Override
         public GroupKey getGroupKey(ResultFile file) {
-            return new ObjectDetectedGroupKey(file);
+            return new DiscoveryKeyUtils.ObjectDetectedGroupKey(file);
         }
 
         @Override
@@ -1375,87 +853,6 @@ public class FileSearch {
     }
 
     /**
-     * Key representing an object detected group
-     */
-    private static class ObjectDetectedGroupKey extends GroupKey {
-
-        private final List<String> objectDetectedNames;
-        private final String objectDetectedNamesString;
-
-        @NbBundle.Messages({
-            "FileSearch.ObjectDetectedGroupKey.noSets=None"})
-        ObjectDetectedGroupKey(ResultFile file) {
-            objectDetectedNames = file.getObjectDetectedNames();
-
-            if (objectDetectedNames.isEmpty()) {
-                objectDetectedNamesString = Bundle.FileSearch_ObjectDetectedGroupKey_noSets();
-            } else {
-                objectDetectedNamesString = String.join(",", objectDetectedNames); // NON-NLS
-            }
-        }
-
-        @Override
-        String getDisplayName() {
-            return getObjectDetectedNamesString();
-        }
-
-        @Override
-        public int compareTo(GroupKey otherGroupKey) {
-            if (otherGroupKey instanceof ObjectDetectedGroupKey) {
-                ObjectDetectedGroupKey otherObjectDetectedGroupKey = (ObjectDetectedGroupKey) otherGroupKey;
-
-                // Put the empty list at the end
-                if (this.getObjectDetectedNames().isEmpty()) {
-                    if (otherObjectDetectedGroupKey.getObjectDetectedNames().isEmpty()) {
-                        return 0;
-                    } else {
-                        return 1;
-                    }
-                } else if (otherObjectDetectedGroupKey.getObjectDetectedNames().isEmpty()) {
-                    return -1;
-                }
-
-                return getObjectDetectedNamesString().compareTo(otherObjectDetectedGroupKey.getObjectDetectedNamesString());
-            } else {
-                return compareClassNames(otherGroupKey);
-            }
-        }
-
-        @Override
-        public boolean equals(Object otherKey) {
-            if (otherKey == this) {
-                return true;
-            }
-
-            if (!(otherKey instanceof ObjectDetectedGroupKey)) {
-                return false;
-            }
-
-            ObjectDetectedGroupKey otherObjectDetectedGroupKey = (ObjectDetectedGroupKey) otherKey;
-            return getObjectDetectedNamesString().equals(otherObjectDetectedGroupKey.getObjectDetectedNamesString());
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(getObjectDetectedNamesString());
-        }
-
-        /**
-         * @return the objectDetectedNames
-         */
-        List<String> getObjectDetectedNames() {
-            return Collections.unmodifiableList(objectDetectedNames);
-        }
-
-        /**
-         * @return the objectDetectedNamesString
-         */
-        String getObjectDetectedNamesString() {
-            return objectDetectedNamesString;
-        }
-    }
-
-    /**
      * Attribute for grouping/sorting by tag name
      */
     static class FileTagAttribute extends AttributeType {
@@ -1482,8 +879,6 @@ public class FileSearch {
             }
         }
     }
-
-
 
     /**
      * Enum for the attribute types that can be used for grouping.
